fixed some more of the failed git submit from before hopefully i found all of those bugs... i suck, mental note remember to pull...
updates to the light toolbar - added some more shadow features in but the whole thing is still a failure, shadows are completely unresponsive on the screen but the settings all work and are persistent, dont know whats missing that is preventing the shadows from showing up. Changed the spotlight to directional light in an attempt to get them to show up, added shadow camera near, far, visible, shadow darkness, shadowmap width and height and bias. Added settings to the renderer in an attempt to get it to work and meshes now have Cast shadows and recieve shadows. No impact on the scene though, very frustrating.

ParticleController was adjusted to allow for multiple particle systems in a scene, each system now keeps a local deep copy of what was in the dat.gui settings to run off of. THE USER MUST HIT REDISTRIBUTE FOR THE SETTINGS TO BE SAVED.  Changed some of the code in sketchCad to load the correct values into dat.gui when switching between them and applied the same scene size functionality that was added in the last push to the systems as well.

This amplified the dizzying factor produced by 1000% at least.