Tentative Grouping push Grouping literally broke everything from the picking tool to the scene explorer lol, added some patches to a bunch of the functions to get them to play nice with the grouping will require testing to make sure it all still works. Pushed this early so you dont yell at me for breaking the save/load again lol
The framework for grouping is in place, controls for making new groups and adding selected meshes to them all seems to be working.
Added a group explorer list that shows what elements are in each group.
Changed how the scene explorer loads in its list apparently scene elements in a group are no longer in the scene object
Changed how the picking function works required a second loop to loop through the groups as well as the scene

Still need to make the gui controller for it and add the functionality to the hot keys.
Still need to add remove from group and delete group buttons as well.
Need to setup the highlighter to highlight the elements in a group when selected.
Need to figure out how to override the single mesh controls with the group mesh controls when a group is selected.